refactor(#211): count관련 eventListener 단일 component화
