+++ backend/application/api/src/main/kotlin/io/raemian/api/cheer/CheeringService.kt
@@ -4,7 +4,7 @@ import io.raemian.api.cheer.controller.request.CheeringRequest
 import io.raemian.api.cheer.controller.request.CheeringSquadPagingRequest
 import io.raemian.api.cheer.controller.response.CheererResponse
 import io.raemian.api.cheer.controller.response.CheeringCountResponse
-import io.raemian.api.cheer.event.CheeringEvent
+import io.raemian.api.event.CheeringEvent
 import io.raemian.api.support.error.CoreApiException
 import io.raemian.api.support.error.ErrorInfo
 import io.raemian.api.support.response.PageResult
