+++ backend/application/api/src/main/kotlin/io/raemian/api/event/CheeringEvent.kt
@@ -1,4 +1,4 @@
-package io.raemian.api.cheer.event
+package io.raemian.api.event
 
 data class CheeringEvent(
     val lifeMapId: Long,
