+++ backend/application/api/src/main/kotlin/io/raemian/api/event/CountEventHandler.kt
@@ -1,21 +1,31 @@
-package io.raemian.api.goal.event
+package io.raemian.api.event
 
+import io.raemian.storage.db.core.cheer.Cheering
+import io.raemian.storage.db.core.cheer.CheeringRepository
 import io.raemian.storage.db.core.goal.GoalRepository
 import io.raemian.storage.db.core.lifemap.LifeMapCount
 import io.raemian.storage.db.core.lifemap.LifeMapCountRepository
 import io.raemian.storage.db.core.lifemap.LifeMapRepository
-import jakarta.transaction.Transactional
 import org.slf4j.LoggerFactory
 import org.springframework.context.event.EventListener
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
-class GoalEventHandler(
+class CountEventHandler(
+    private val cheeringRepository: CheeringRepository,
     private val lifeMapCountRepository: LifeMapCountRepository,
-    private val goalRepository: GoalRepository,
-    private val lifeMapRepository: LifeMapRepository,
 ) {
     private val log = LoggerFactory.getLogger(javaClass)
+
+    @Transactional
+    @EventListener
+    fun addCheeringCount(cheeringEvent: CheeringEvent) {
+        val cheering = cheeringRepository.findByLifeMapIdForUpdate(cheeringEvent.lifeMapId)
+            ?: Cheering(0, cheeringEvent.lifeMapId)
+
+        cheeringRepository.save(cheering.addCount())
+    }
 
     @Transactional
     @EventListener
