+++ backend/application/api/src/main/kotlin/io/raemian/api/goal/GoalService.kt
@@ -5,7 +5,7 @@ import io.raemian.api.goal.controller.request.UpdateGoalRequest
 import io.raemian.api.goal.controller.response.CreateGoalResponse
 import io.raemian.api.goal.controller.response.GoalResponse
 import io.raemian.api.goal.domain.GoalExploreDTO
-import io.raemian.api.goal.event.CreateGoalEvent
+import io.raemian.api.event.CreateGoalEvent
 import io.raemian.api.sticker.StickerService
 import io.raemian.api.support.RaemianLocalDate
 import io.raemian.api.support.error.MaxGoalCountExceededException
